Reorganize code to help with readability/maintainability
