+++ romanesco/utils.py
@@ -1,0 +1,39 @@
+import functools
+
+
+def toposort(data):
+    """
+    General-purpose topological sort function. Dependencies are expressed as a
+    dictionary whose keys are items and whose values are a set of dependent
+    items. Output is a list of sets in topological order. This is a generator
+    function that returns a sequence of sets in topological order.
+
+    :param data: The dependency information.
+    :type data: dict
+    :returns: Yields a list of sorted sets representing the sorted order.
+    """
+    if not data:
+        return
+
+    # Ignore self dependencies.
+    for k, v in data.items():
+        v.discard(k)
+
+    # Find all items that don't depend on anything.
+    extra = functools.reduce(
+        set.union, data.itervalues()) - set(data.iterkeys())
+    # Add empty dependences where needed
+    data.update({item: set() for item in extra})
+
+    # Perform the toposort.
+    while True:
+        ordered = set(item for item, dep in data.iteritems() if not dep)
+        if not ordered:
+            break
+        yield ordered
+        data = {item: (dep - ordered)
+                for item, dep in data.iteritems() if item not in ordered}
+    # Detect any cycles in the dependency graph.
+    if data:
+        raise Exception('Cyclic dependencies detected:\n%s' % '\n'.join(
+                        repr(x) for x in data.iteritems()))
